requirements.txt
streamlit
requests
python-dotenv
PyPDF2
python-docx
